signup works, login works with username as undefined
