+++ server/controllers/user-controller.js
@@ -40,6 +40,10 @@ const userController = {
   // {body} is destructured req.body
   async login({ body }, res) {
     const user = await User.findOne({ $or: [{ username: body.username }, { email: body.email }] }).select('-__v');
+
+    console.log(user)
+    console.log('=====username ', user.username);
+
     if (!user) {
       return res.status(400).json({ message: "Can't find this user" });
     }
